requirements_ml.txt
# VisioNova Python 3.10 Environment (ML Models with CUDA)
#
# This environment is specifically for GPU-accelerated ML models.
# Python 3.13+ is NOT compatible with PyTorch CUDA builds.
#
# Installation:
#   1. Create environment: py -3.10 -m venv .venv310
#   2. Activate: .venv310\Scripts\Activate.ps1
#   3. Install: pip install -r requirements_ml.txt

# Core ML frameworks (CUDA 12.1)
torch==2.5.1+cu121
torchvision==0.20.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# HuggingFace ecosystem
transformers==5.1.0
safetensors==0.7.0
tokenizers==0.22.2
huggingface-hub==1.4.1

# Computer vision
opencv-python==4.13.0.92
pillow==12.0.0

# Scientific computing
numpy==2.2.6
scipy==1.15.3

# Flask backend
flask==3.1.2
flask-cors==6.0.2
requests==2.32.5

# AI services
groq==1.0.0
python-dotenv==1.2.1

# Web scraping (for fact-checking)
beautifulsoup4==4.14.3

# Data validation
pydantic==2.12.5
